Incorporating suggestions by Nhu.
Added:
(1) Justification as to the selection of the SVAR model over alternative methodologies.
(2) Elaborated on how the A matrix can be interpreted, such as how entries in A denote the immediate impact of shocks in one variable on others within the same time period.
(3) Tighten up the explanation of the Granger causality tests to make it more concise while still highlighting how they add depth to the analysis.